fix(jade-garden): inferred type
