+++ packages/core/src/types.ts
@@ -30,6 +30,20 @@ type Attribute<T> =
 type ClassArray = ClassValue[];
 
 /**
+ * Represents the `class` and `className` props for `cva` and `sva`.
+ * Ensures that only one of `class` or `className` is present.
+ */
+export type ClassProp =
+  | {
+      class?: ClassValue;
+      className?: never;
+    }
+  | {
+      class?: never;
+      className?: ClassValue;
+    };
+
+/**
  * Represents the minimum structure to work with class names.
  * Fully compatible input for `jade-garden/class-utils` functions.
  */
@@ -82,6 +96,45 @@ export type CreateOptions = {
 };
 
 /**
+ * **FOR LIBRARY AUTHORS**
+ *
+ * Add JSDoc and CSS comments to components when generating with `unplugin-jade-garden`.
+ */
+export type MetaConfig = {
+  /**
+   * Adds a `deprecated` tag.
+   *
+   * Adds a description if `deprecated` is a string.
+   *
+   * @see https://jsdoc3.vercel.app/tags/deprecated
+   */
+  deprecated?: boolean | string;
+
+  /**
+   * Adds a `description` tag.
+   *
+   * @see https://jsdoc3.vercel.app/tags/description
+   */
+  description?: string;
+
+  /**
+   * Adds a `name` tag.
+   *
+   * `name` should be the same as `name` in style configuration.
+   *
+   * @see https://jsdoc3.vercel.app/tags/name
+   */
+  name?: string;
+
+  /**
+   * Adds a `see` tag.
+   *
+   * @see https://jsdoc3.vercel.app/tags/see
+   */
+  see?: string;
+};
+
+/**
  * The merge function signature for `createCVA` and `createSVA`.
  * @param {...ClassValue[]} inputs - A variadic number of arguments of type {@link ClassValue}.
  * @returns {string} A single string of merged class names.
@@ -95,6 +148,12 @@ export type MergeFn = (...classes: ClassValue[]) => string;
  * @returns {T extends undefined ? never : T} The type with undefined removed.
  */
 type OmitUndefined<T> = T extends undefined ? never : T;
+
+/**
+ * Converts a string literal "true" or "false" to its boolean type.
+ * @template T - The string literal type, e.g., "true" | "false".
+ */
+export type StringToBoolean<T> = T extends "true" | "false" ? boolean : T;
 
 /**
  * Provides type safety for the `data` prop within the `traits` function.
